% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_durations.R
\name{calculate_durations}
\alias{calculate_durations}
\title{Calculate durations}
\usage{
calculate_durations(df, probs = c(10, 25, 75, 90))
}
\arguments{
\item{df}{a data frame with a "lake" and a "level" column}

\item{probs}{a vector with all exceedance probabilities to calculate.
Defaults to c(10, 25, 75, 90).}
}
\value{
durations, a data frame with the following columns:
\item{lake}{name of lake}
\item{variable}{exceedance probability, e.g., "10", "25", "75", or "90"}
\item{value}{one count of number of months levels were consecutively above or
below the given exceedance probability}
}
\description{
Given a data frame with a "lake", "date", and "level" columns as well as a
vector of desired exceedance probabilities, calculates consecutive months
at/above (for probabilities <= 50\%) or at/below (for probabilities > 50\%) the
exceedance levels.
}
