% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_metrics.R
\name{calculate_exceedances}
\alias{calculate_exceedances}
\title{Calculate exceedance levels}
\usage{
calculate_exceedances(df)
}
\arguments{
\item{df}{a data frame with a "lake" and a "level" column}
}
\value{
vals, a data frame with the names of all lakes and corresponding 10\%,
25\%, 50\%, 75\%, and 90\% exceedance levels
}
\description{
Given a data frame with a "lake" and "level" column, calculates the 10\%, 25\%,
50\%, 75\%, and 90\% exceedance levels
}
