% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fit.R
\name{calculate_fit}
\alias{calculate_fit}
\title{Calculate goodness of fit metrics}
\usage{
calculate_fit(obs, sim, metric)
}
\arguments{
\item{obs}{a single observed value}

\item{sim}{a vector with all simulated values}

\item{metric}{a string indicating which goodness-of-fit metric to use.
Options include "PBIAS", "CV", or "RMSE".}
}
\value{
fit, a vector corresponding to "sim" with the goodness of fit metrics
for each observation.
}
\description{
Given an observed value, a list of simulated values, and the desired metric,
calculates the goodness of fit.
}
