% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_levels_at_probs.R
\name{calculate_levels_at_probs}
\alias{calculate_levels_at_probs}
\title{Calculate lake level at given exceedance probability}
\usage{
calculate_levels_at_probs(
  ranked,
  probs = c(10, 50, 90),
  lakes = c("Pleasant", "Long", "Plainfield")
)
}
\arguments{
\item{ranked}{a data frame with the following columns
\itemize{
\item \strong{lake:} the name of the lake ("Pleasant", "Long", or "Plainfield)
\item \strong{season:} the season of the prediction
\item \strong{date:} the month and year of prediction in datetime (POSIXct) format
\item \strong{level:} the predicted lake level (meters above mean sea level)
\item \strong{prob:} the probability of a lake level that equals or exceeds the
predicted level (percent)
}}

\item{probs}{the exceedance probabilities of interest (percent), defaults to
10, 50, and 90.}

\item{lakes}{the names of the lakes in the order you would like them
displayed in plots, etc. Defaults to c("Pleasant", "Long",
"Plainfield")}
}
\value{
df, a data frame with the following columns:
\item{lake}{the name of the lake ("Pleasant", "Long", or "Plainfield)}
\item{season}{the season (e.g., annual, spring, summer, winter, or fall)}
\item{prob}{the exceedance probability}
\item{level}{the lake level associated with the given exceedance probability
for the given lake and season}
}
\description{
Given a dataset with lake levels and exceedance probabilities, this function
interpolates the lake level associated with the provided exceedance
probability for each lake and season in the dataset.
}
