% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_metrics.R
\name{calculate_metrics}
\alias{calculate_metrics}
\title{Calculate hydrologic metrics evaluated in the CSLS}
\usage{
calculate_metrics(
  df = CSLSlevels::hist_levels,
  metrics = c("median_level", "cv_level", "exceedance_level", "exceedance_range",
    "depart_median", "median_dur", "median_dur_seas", "cv_dur", "cv_dur_seas", "num_dur",
    "num_dur_seas", "num_2yr", "num_2yr_seas", "num_dur_decade", "num_dur_decade_seas",
    "num_2yr_decade", "num_2yr_decade_seas", "median_rise_rate", "cv_rise_rate",
    "median_fall_rate", "cv_fall_rate", "fast_rise", "fast_fall", "good_spawning")
)
}
\arguments{
\item{df}{data frame to use, defaults "CSLSlevels::hist_levels". Must include
columns for "date" (must be POSIXct), "lake" (must be factor), and
one with the lake levels (named in "col_name" argument).}

\item{metrics}{a list of which metrics to use. Defaults to all of them
c("median_level", "cv_level", "exceedance_level",
"exceedance_range", "median_dur", "cv_dur",
"median_rise_rate", "cv_rise_rate", "median_fall_rate",
"cv_fall_rate").}
}
\value{
summary, a data frame with the following columns:
\item{lake}{name of lake, character}
\item{metric}{name of metric, corresponds with values in inputted "metrics"
argument, character}
\item{variable}{identifier for different types of metric. For example, the
metric "median_rise_rate" will have variables of "1", "3",
and "12" associated with it indicating the median rate over
1 month, 3 months, and 12 months, character}
\item{value}{calculated value of the metric, numeric}
}
\description{
This function calculates hydrologic metrics describing the magnitude,
frequency, duration, rate of change, and timing of lake levels. Specific
metrics calculated include:
\itemize{
\item \strong{median_level:} median monthly and overall lake levels
\item \strong{cv_level:} coefficient of variation for monthly and overall lake levels
\item \strong{exceedance_level:} lake levels corresponding to 10\%, 25\%, 50\%, 75\%,
and 90\% exceedance probabilities
\item \strong{exceedance_range:} range between 25\% and 75\% exceedance probabilities
and 10\% and 90\% exceedance probabilities for all
levels and annual mean levels.
\item \strong{median_dur:} median duration at/above the 10\% and 25\% exceedance
probabilities or at/below the 75\% and 90\% exceedance
probabilities.
\item \strong{cv_dur:} coefficient of vatiation for durations at/above the 10\% and
25\% exceedance probabilities and at/below the 75\% and 90\%
exceedance probabilities.
\item \strong{median_rise_rate:} median lake level rise over 1 month, 3 months
(seasonal), and 12 months (annual).
\item \strong{cv_rise_rate:} coefficient of variation in lake level rise over 1
month, 3 months (seasonal), and 12 months (annual).
\item \strong{median_fall_rate:} median lake level rise over 1 month, 3 months
(seasonal), and 12 months (annual).
\item \strong{cv_fall_rate:} coefficient of variation in lake level rise over 1
month, 3 months (seasonal), and 12 months (annual).
}
}
