% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_probs_of_levels.R
\name{calculate_probs_of_levels}
\alias{calculate_probs_of_levels}
\title{Calculate exceedance probabilities given lake level timeseries}
\usage{
calculate_probs_of_levels(
  levels,
  start_yr = 1981,
  end_yr = 2018,
  seasons = list(annual = seq(1, 12, 1)),
  nyears = NULL
)
}
\arguments{
\item{levels}{a data frame with the following columns:
\itemize{
\item \strong{lake:} the name of the lake ("Pleasant", "Long", or "Plainfield")
\item \strong{date:} the month an year of prediction in datetime (POSIXct) format
\item \strong{level_pred:} predicted lake level, in meters above mean sea level
}}

\item{start_yr}{year to start considering lake level data (inclusive),
defaults to 1981}

\item{end_yr}{year to stop considering lake level data (inclusive),
defaults to 2018}

\item{seasons}{a list with the season names and the month numbers associated
with each season. For example, list(summer = c(6, 7, 8), fall
= c(9, 10, 11), winter = c(12, 1, 2), spring = c(3, 4, 5)).
Defaults to list(annual = seq(1, 12, 1)).}

\item{nyears}{number of years to randomly sample for observations from
between start_yr and end_yr. Defaults to NULL to use all years
between start_yr and end_yr (inclusive).}
}
\value{
ranked, a data frame with the following columns:
\item{lake}{the name of the lake ("Pleasant", "Long", or "Plainfield)}
\item{season}{the season of the prediction}
\item{date}{the month and year of prediction in datetime (POSIXct) format}
\item{level}{the predicted lake level (meters above mean sea level)}
\item{prob}{the probability of a lake level that equals or exceeds the
predicted level (percent)}
}
\description{
Given a timeseries of lake levels and a specified time period, this function
ranks lake levels by lake and season and calculates the lake level exceedance
probabilities.
}
