% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_rates.R
\name{calculate_rates}
\alias{calculate_rates}
\title{Calculate rates of change}
\usage{
calculate_rates(df, months = c(1, 3, 12))
}
\arguments{
\item{df}{a data frame with a "lake" and a "level" column}

\item{months}{a vector with all lag times (in months) to consider, defaults
to c(1, 3, 12).}
}
\value{
rates, a data frame with the following columns:
\item{lake}{name of lake}
\item{variable}{number of months lagged (e.g., 1, 3, 12)}
\item{value}{change in lake levels over given time period}
}
\description{
Given a data frame with a "lake", "date", and "level" columns as well as a
vector of desired lag (e.g., 1 month, 3 months, 12 months), calculates the
change in lake levels over each time period.
}
