% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_spawning.R
\name{calculate_spawning}
\alias{calculate_spawning}
\title{Calculate good spawning years}
\usage{
calculate_spawning(df, growing_months = c(4, 5, 6, 7, 8, 9, 10))
}
\arguments{
\item{df}{a data frame with a "lake" and a "level" column}

\item{growing_months}{months associated with growing season. Defaults to
c(5, 6, 7, 8, 9, 10).}
}
\value{
good_spawning, a data frame with the lakes, years, and whether spring
levels were sufficiently high, summer levels were sufficiently steady, and
overall the year was good for pike spawning.
}
\description{
Given a data frame with a "lake", "level", "year", and "month" columns,
calculates whether a year is good (TRUE) or bad (FALSE) for pike spawning.
}
