% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_season.R
\name{classify_season}
\alias{classify_season}
\title{Classify season of observations}
\usage{
classify_season(
  df,
  seasons = list(summer = c(6, 7, 8), fall = c(9, 10, 11), winter = c(12, 1, 2), spring
    = c(3, 4, 5))
)
}
\arguments{
\item{df}{the input data frame, with a column for "date" that has the
observation datetime in POSIXct.}

\item{seasons}{a list of the names of each season and the vectors of which
months fall into each season (e.g., list(summer = c(6, 7, 8),
fall = c(9, 10, 11))).}
}
\value{
df, the input data frame with a column added for the season of the
observation.
}
\description{
Given a timeseries of observations and a list of which month falls in which
season, classify what season each observation falls into.
}
