% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_veg_transitions.R
\name{classify_veg_transitions}
\alias{classify_veg_transitions}
\title{Classify vegetation transition depths}
\usage{
classify_veg_transitions(
  lake,
  deepest = data.frame(upland = -1, inland_beach = 1.6, emergent = 7, floating = 10),
  shallowest = data.frame(floating = 5, submergent = 1.6, emergent = -1, inland_beach =
    -1.6),
  contour_interval = 0.1,
  depth_to_meters = TRUE
)
}
\arguments{
\item{lake}{name of lake to analyze, e.g., "Long". Must be present in both
CSLSlevels::csls_levels and CSLSdata::lake_levels.}

\item{deepest}{list with deepest depth each vegetation class can tolerate
(ft). If depth_to_meters is TRUE, these will be converted to
meters. If deepest not provided for "submergent", assumed to
be the max depth of the lake.}

\item{shallowest}{list with shallowest depth each vegetation class can
tolerate (ft). If depth_to_meters is TRUE, these will be
converted to meters. If shallowest not provided for
"upland", assumed to be the negative max depth of the lake.}

\item{contour_interval}{interval of contours to use when creating
relationship, same units as desired transition depths
data frame. Defaults to 0.1.}

\item{depth_to_meters}{defaults to TRUE, indicating transition depths will be
in meters and deepest and shallowest should be
converted from feet to meters.
.}
}
\value{
transitions, a data frame with the following columns:
\item{depth}{depth of transition point (meters)}
\item{class}{name of vegetation classes present at/deeper than this depth}
}
\description{
Given the name of the lake of interest and shallow/deep limits for vegetation
classes, this defines transition depths between classes of vegetation. It is
not necessary to provide the shallowest depth of "upland" or the deepest
depth for "submergent" if these correspond with the maximum depth of the
lake. If desire depths in ft, be sure to set depth_to_meters to FALSE and
provice countour_interval in feet. Assumes deepest and shallowest depths are
always provided in ft.
}
