% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_lake_contours.R
\name{draw_lake_contours}
\alias{draw_lake_contours}
\title{Draw contours for a given lake}
\usage{
draw_lake_contours(ranked_levels, lake, season)
}
\arguments{
\item{ranked_levels}{a data frame with the following columns:
\itemize{
\item{lake}{name of the lake, e.g., "Long Lake"}
\item{season}{name of season}
\item{prob}{exceedance probability (percent)}
\item{level}{lake level associated with that exceedance probability for the
given lake and season}
}}

\item{lake}{name of lake to analyze, e.g., "Long"}

\item{season}{name of season ot analyze, e.g., "summer"}
}
\value{
plot_obj, a plot of the lake outline for each exceedance probability
in the dataset for the given lake and season.
}
\description{
Given a dataset with lake levels and exceedance probabilities, this loads the
raster associated with the lake and draws contours at the desired levels.
}
