% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establish_lake_extents.R
\name{establish_lake_extents}
\alias{establish_lake_extents}
\title{Establish lake extents}
\usage{
establish_lake_extents(lake, contour_interval = 0.01, min_deepest = 0)
}
\arguments{
\item{lake}{name of lake to analyze, e.g., "Long". Must be present in both
CSLSlevels::csls_levels and CSLSdata::lake_levels.}

\item{contour_interval}{interval of contours to use when creating
relationship, or precision of lake level (m).
Defaults to 0.01.}

\item{min_deepest}{minimum value of deepest vegetation classes (m). Defaults
to zero so does not account for vegetation classes deeper
than the max elevation of lake.}
}
\value{
elev, a list with the minimum and maxiumum elevation of the lake.
}
\description{
Defines minimum and maxiumum lake elevations based on deepest point of lake,
maximum observed level, and desired contour interval.
}
