% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_lake_level.R
\name{find_lake_level}
\alias{find_lake_level}
\title{Find lake level at given exceedance probability}
\usage{
find_lake_level(ranked, year = NA, type = "forward", window = 0, prob = 50)
}
\arguments{
\item{ranked}{a data frame with the lake name (lake), lake levels
(level) and exceedance probabilities (prob).}

\item{year}{the end year of the analysis window used to calculate lake level
exceedance probabilities, defaults to NA (optional).}

\item{type}{type of looping done, defaults to "forward". Also "backward".}

\item{window}{length of window of analysis. Defaults to zero to indicate used
entire record.}

\item{prob}{the exceedance probability of interest (percent), defaults to 50.}
}
\value{
df, a data frame which track the provided end year of the analysis
window (end_year) and desired exceedance probability (prob) as well as
the corresponding lake level (level) for each lake (lake).
}
\description{
Given a dataset with lake levels and exceedance probabilities, this function
interpolates the lake level associated with the provided exceedance
probability for each lake in the dataset.
}
