% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_lake_level_seasonal.R
\name{find_lake_level_seasonal}
\alias{find_lake_level_seasonal}
\title{Find lake level at given exceedance probabilities for each season}
\usage{
find_lake_level_seasonal(
  ranked,
  probs = c(10, 50, 90),
  seasons = c("winter", "fall", "spring", "summer")
)
}
\arguments{
\item{ranked}{a data frame with the lake name (lake), lake levels
(level) and exceedance probabilities (prob) for each season
(season).}

\item{probs}{the exceedance probabilities of interest (percent), defaults to
10, 50, and 90.}

\item{seasons}{name of seasons, defaults to "winter", "fall", "spring", and
"summer".}
}
\value{
df, a data frame with the following columns:
\item{lake}{name of the lake, e.g., "Long Lake"}
\item{season}{name of season}
\item{prob}{exceedance probability (percent)}
\item{level}{lake level associated with that exceedance probability for the
given lake and season}
}
\description{
Given a dataset with lake levels and exceedance probabilities, this function
interpolates the lake level associated with the provided exceedance
probability for each lake in the dataset.
}
