% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_level_functions.R
\name{find_multiple_levels}
\alias{find_multiple_levels}
\title{Find lake level at given exceedance probability for multiple windows of analyses}
\usage{
find_multiple_levels(levels, years, probs = c(10, 50, 90),
  types = "forward", windows = 0)
}
\arguments{
\item{levels}{a data frame with the lake name (lake_name), month and year of
prediction (obs_mo), and predicted lake level (lake_lev_pred)}

\item{years}{the end years of the analysis window used to calculate lake level
exceedance probabilities.}

\item{probs}{the exceedance probabilities of interest (percent), defaults to
c(10, 50, 90)}

\item{types}{type of looping to do, defaults to "forward". Also "backward".}

\item{windows}{length of window of analysis. Defaults to zero to indicate use
entire record.}
}
\value{
df, a data frame which track the provided end year of the analysis
        window (end_year) and desired exceedance probability (prob) as well as
        the corresponding lake level (level) for each lake (lake).
}
\description{
Given a dataset with lake levels and exceedance probabilities, this function
interpolates the lake level associated with the provided exceedance
probability for each lake in the dataset.
}
