% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hi_lo_ann.R
\name{hi_lo_ann}
\alias{hi_lo_ann}
\title{CV and mean of number of annual occurences of extreme levels}
\usage{
hi_lo_ann(
  lakeName,
  exceeds = c(0.1, 0.25, 0.75, 0.9),
  startDate = "1905-01-01",
  endDate = "2019-12-01"
)
}
\arguments{
\item{lakeName}{Name of the lake of interest without "lake", e.g.
"Plainfield", "Long", or "Pleasant"}

\item{exceeds}{The exceedance probabilities of interest, as a vector of
values between 0 and 1}

\item{startDate}{Specify a range of dates by which to filter lake level
predictions}

\item{endDate}{Specify a range of dates by which to filter lake level
predictions}
}
\value{
a list of the following:
\item{result}{A data frame of one column, with mean and cv for each
exceedance supplied, with appropriate row names}
\item{df}{The data frame generated by hi_lo_duration() that was used by this
function}
}
\description{
Given the cslsLevels data frame, calculate the percentage of years in which
the lake spends at least one month beyond given exceedance probabilities
}
