% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hi_lo_duration.R
\name{hi_lo_duration}
\alias{hi_lo_duration}
\title{Calculate duration of extreme lake levels}
\usage{
hi_lo_duration(
  lakeName,
  exceeds = c(0.1, 0.25, 0.75, 0.9),
  elevs = NULL,
  pred_colname = "level_pred",
  date_colname = "date",
  lake_colname = "lake",
  startDate = "1905-01-01",
  endDate = "2019-12-01"
)
}
\arguments{
\item{lakeName}{Name of the lake of interest without "lake", e.g. "Plainfield", "Long", or "Pleasant"}

\item{exceeds}{The exceedance probabilities of interest, as a vector of values between 0 and 1}

\item{elevs}{The elevations of interest. If this argument is supplied, "exceeds" is ignored.}

\item{pred_colname}{The column of CSLSlevels::csls_levels to look for predicted lake levels (in case that changes)}

\item{date_colname}{Same as above, but for date}

\item{lake_colname}{Same as above, but for lake}

\item{startDate}{Enter a different start date for the analysis, or leave blank to use the entire sequence of
predicted lake levels (1905- 2019)}

\item{endDate}{Enter a different end date for the analysis, or leave blank to use the entire sequence of
predicted lake levels (1905- 2019)}
}
\value{
a list of the following:
\item{ll_hi_lo} {The lake level predictions and dates used in the calculation, with extra columns added for
each exceedance or elevation of interest. One column with a 1 for every monthly prediciton more extreme than the exceedance or
elevation of interest, and one with number of meters above or below the exceedance or elevation of interest}
\item{dur_counts} {A data frame with a row for every sequence of months above or below the exceedance or elevation of interest}
\item{mmRes} {The same as above except with cumulative meters above or below the exceedance or elevation of interest for
consecutive months}
\item{levels} {The median predicted level and the elevations of interest used to perform the calculations.}
}
\description{
Calculate how long a lake remains above or below, given either exceedance probabilities or elevations in meters above
sea level.
}
