% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kennard_metrics.R
\name{Kennard_metrics}
\alias{Kennard_metrics}
\title{Calculate a lot of metrics found in appendix to Kennard (2009)}
\usage{
Kennard_metrics(
  LakeLevels = CSLSlevels::csls_levels,
  startDate = NULL,
  endDate = NULL,
  LakeLevelColumn = "level_pred",
  metrics = c("meanMonth", "cvMonth", "hiLevs", "ARIs", "hiloAnn", "hiloDur",
    "hiloDurMM", "hiloCV", "meanRate", "cvRate", "mn_ann_Range")
)
}
\arguments{
\item{LakeLevels}{Data frame to use, "csls_levels" by default.}

\item{startDate}{Earliest date to use in calculations in YYY-MM-DD format.
Null results in the entire range of dates being used.}

\item{endDate}{Latest data to use in calculations in YYYY-MM-DD format.}

\item{LakeLevelColumn}{Column of data frame to use. Default refers to column
in "csls_levels". Will likely not need to be changed.}

\item{metrics}{A list of named, logical values corresponding to the metrics
to be used. By default all metrics will be calculated. The
object "which_Kennard_metrics" can be used here to easily
exclude some metrics.}
}
\value{
A list of metric results.
}
\description{
Uses the CSLSlevels data frame and calculates metrics related to length,
frequency and coeffiecient of variation of extreme lake levels.
}
\examples{
  Kennard_metrics()
  Kennard_metrics(metrics = c("meanMonth", "hiLevs"))

}
