% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lake_raster.R
\docType{data}
\name{lake_raster}
\alias{lake_raster}
\title{Dataset: CSLS lake elevation rasters}
\format{
A list with three rasters, one for each lake.
\describe{
\item{Pleasant}{elevation raster for Pleasant Lake}
\item{Long}{elevation raster for Long Lake}
\item{Plainfield}{elevation raster for Plainfield Lake}
}
}
\usage{
data(lake_raster)
}
\description{
Rasters with elevation data around the CSLS lakes. Rasters clipped to highest
recorded lake level plus a 20 meter buffer.
}
\details{
TIF files are loaded and save as rasters with the \code{save_lake_rasters.R}
script in the \code{data-raw} subdirectory.
}
\keyword{datasets}
