% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_metrics.R
\name{melt_and_add_metric}
\alias{melt_and_add_metric}
\title{Calculate exceedance levels}
\usage{
melt_and_add_metric(vals, metric_name = NULL, prefix = NULL)
}
\arguments{
\item{vals}{a data frame with a "lake" and a "level" column}

\item{metric_name}{name of metric to append in a "metric" column. Defaults
to NULL to not include.}

\item{prefix}{prefix of melted column names to remove from entries in melted
"variable" column. Defaults to null to keep names as-is.}
}
\value{
vals, a data frame with the columns:
\item{lake}{name of the lake}
\item{metric}{name of metric passed to function}
\item{variable}{name of variations on the metric, columnames in df}
\item{value}{value of each metric for each lake}
}
\description{
Given a data frame with a "lake" and "level" column, calculates the 10\%, 25\%,
50\%, 75\%, and 90\% exceedance levels
}
