% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo.R
\docType{data}
\name{monte_carlo}
\alias{monte_carlo}
\title{Dataset: monte carlo sampling of lake levels 1905-2018}
\format{A data frame with the 10, 50, and 90 percent exceedance probability
lake level under different random samples of years for CSLS lakes.
\describe{
\item{lake}{lake name, "Long", "Plainfield", "Pleasant", or "Devils"}
\item{date}{month and year of observation, POSIXct}
\item{level}{lake level (mamsl) on that date}
\item{sim}{simulation number}
\item{nyears}{number of years sampled}
}}
\usage{
data(monte_carlo)
}
\description{
Lake levels associated with an increasing number of randomly sampled years
for each lake.
}
\keyword{datasets}
