% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo_2015.R
\docType{data}
\name{monte_carlo_2015}
\alias{monte_carlo_2015}
\title{Monte Carlo 2015 results}
\format{A data frame with the 10, 50, and 90 percent exceedance probability
  lake level under different random samples of years for CSLS lakes.
\describe{
  \item{year}{number of years randomly sampled}
  \item{type}{"random"}
  \item{window}{number of monte carlo run}
  \item{prob}{exceedance probability}
  \item{lake}{Lake name, "Long Lake", "Plainfield Lake", or "Pleasant Lake"}
  \item{level}{Lake level (mamsl)}
}}
\usage{
data(monte_carlo_2015)
}
\description{
Lake levels at specific exceedance probabilities from different years of
analysis.
}
\keyword{datasets}
