% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo_functions.R
\name{monte_carlo_exceedances}
\alias{monte_carlo_exceedances}
\title{Monte Carlo Analysis of Lake Levels}
\usage{
monte_carlo_exceedances(levels, ts_start = 1905, ts_end = 2018,
  nreplicates = 500, probs = c(10, 50, 90))
}
\arguments{
\item{levels}{a data frame with the lake name (lake_name), month and year of
prediction (obs_mo), and predicted lake level (lake_lev_pred)}

\item{ts_start}{year to start considering lake level data (inclusive),
defaults to 1905}

\item{ts_end}{year to stop considering lake level data (inclusive), defaults
to 2018.}

\item{nreplicates}{number of monte carlo replicates to perform, defaults to
500.}

\item{probs}{probabilities to highlight with dashed vertical line, defaults
to 10, 50, and 90 percent.}
}
\value{
monte_carlo, a data frame with the number of years used (year), note
  that these are random windows (type), monte carlo loop number (window),
  exceedance probability (prob), lake name (lake), and lake level (level).
}
\description{
This function loops thorugh a Monte Carlo analysis of increasing the size of
the window of analysis by randomly sampling different years.
}
