% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo_metrics.R
\docType{data}
\name{monte_carlo_metrics}
\alias{monte_carlo_metrics}
\title{Dataset: hydrologic metrics calculated with random samples of lake levels}
\format{
A data frame with the following columns.
\describe{
\item{lake}{name of lake, e.g., Pleasant, Long, Plainfield, Devils}
\item{metric}{name of hydrologic metrics, e.g. median_level, cv_rise_rate}
\item{variable}{name of variation on metrics, e.g. 10 for 10\% exceedance
level}
\item{sim}{value of hydrologic metric calculated using randomly sampled levels}
\item{nsim}{id number for this set of randomly sampled levels}
\item{nyear}{number of years randomly sampled in this simulation}
\item{obs}{value of hydrologic metric calculated using all valid levels in
timeseries}
}
}
\usage{
data(monte_carlo_metrics)
}
\description{
Hydrologic metrics calculated using randomly sampled lake levels.
}
\keyword{datasets}
