% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bathymetry.R
\name{plot_bathymetry}
\alias{plot_bathymetry}
\title{Draw depth bathymetry contours for a given lake}
\usage{
plot_bathymetry(
  lake,
  max_levels_ft = data.frame(lake = c("Pleasant", "Long", "Plainfield"), max_ft =
    c(984, 1105, 1106), max_level = c(983.5, 1105.1, 1104.7)),
  fill = list(Pleasant = list(limits = c(-5, 25), breaks = c(20, 15, 10, 5, 0, -5),
    labels = c("20", "15", "10", "5", "0", "+5")), Long = list(limits = c(-7.5, 7.5),
    breaks = c(6, 3, 0, -3, -6), labels = c("6", "3", "0", "+3", "+6")), Plainfield =
    list(limits = c(-8.5, 8.5), breaks = c(8, 4, 0, -4, -8), labels = c("8", "4", "0",
    "+4", "+8"))),
  lakes = c("Pleasant", "Long", "Plainfield"),
  text_size = 10
)
}
\arguments{
\item{lake}{name of lake to analyze, e.g., "Long"}

\item{max_levels_ft}{maximum observed lake levels and max contour lake levels
at each lake.}

\item{fill}{breaks, limits, and labels for fill colors at each lake}

\item{lakes}{vector will all lake names}

\item{text_size}{size of text on plot}
}
\value{
plot_obj, a plot with the depth contours of the lake.
}
\description{
Draw depth bathymetry contours for a given lake
}
