% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_duration.R
\name{plot_duration}
\alias{plot_duration}
\title{Plot duration above/below exceedance levels}
\usage{
plot_duration(
  df,
  probs = 90,
  show_median = TRUE,
  show_labels = TRUE,
  max_months = NULL,
  title = "",
  text_size = 12,
  vline_color = "#c00000",
  hist_color = "grey80",
  density_color = "#c00000",
  line_size = 1,
  bin_size = 1,
  lakes = c("Pleasant", "Long", "Plainfield")
)
}
\arguments{
\item{df}{a data frame with columns for "lake", "date", "level_obs", and
"level_pred" that matches desired time series to analyze}

\item{probs}{exceedance probabilities to analyze, defaults to 90}

\item{show_median}{logical defaults to FALSE. If true, plots vertical line at
median number of months.}

\item{show_labels}{logical defaults to TRUE to label median duration (if displayed)}

\item{max_months}{default to NULL, otherwise used to set max}

\item{title}{string to use for title of plot, defaults to "".}

\item{text_size}{size of text, defaults to 12}

\item{vline_color}{color of vertical line, defaults to a red ("#c00000")}

\item{hist_color}{color of histogram bins, defaults to "grey80"}

\item{density_color}{color of density curve with alpha = 0.2, defaults to a
red ("#c00000")}

\item{line_size}{line size, defaults to 1}

\item{bin_size}{width of bins, defaults to 2 months.}

\item{lakes}{vector of lakes to include in plot. Defaults to c("Pleasant",
"Long", "Plainfield")}
}
\value{
plot_obj, a plot with the imputed and observed lake levels.
}
\description{
Given a data frame with lake levels for a single lake or multiple lakes, plot
the duration of . If data frame has multiple lakes,
will display as facet plots. Options to add yintercepts. Also option to add
second axis with lake depth, if only plotting one lake.
}
