% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_levels_functions.R
\name{plot_exceedance_curve}
\alias{plot_exceedance_curve}
\title{Plot lake levels exceedance curves}
\usage{
plot_exceedance_curve(df, probs = c(10, 50, 90), text_size)
}
\arguments{
\item{df}{data frame to plot with the exceedance probability (prob), lake
level (level), end year of analysis (end_year), and lake name
(lake_name)}

\item{probs}{probabilities to highlight with dashed vertical line, defaults
to 10, 50, and 90 percent.}

\item{text_size}{size of text in plots}
}
\value{
plot_obj, the plot.
}
\description{
This function plots the lake level at three predetermined exceedance
probabilities (e.g., 10, 50, 90 percent) for a range of analysis end years at all
lakes.
}
