% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_frequency.R
\name{plot_frequency}
\alias{plot_frequency}
\title{Plot distribution of inputed lake levels}
\usage{
plot_frequency(
  df,
  convert_to_ft = TRUE,
  xintercept = NULL,
  probs = NULL,
  title = "",
  text_size = 12,
  vline_color = "#c00000",
  hist_color = "grey80",
  density_color = "#c00000",
  line_size = 1,
  pfl_is_long = TRUE,
  labels = TRUE,
  lakes = c("Pleasant", "Long", "Plainfield")
)
}
\arguments{
\item{df}{a data frame with columns for "lake", "date", "level_obs", and
"level_pred".}

\item{convert_to_ft}{defaults to TRUE to convert lake levels from meters to ft}

\item{xintercept}{either a single value (if only one lake) or a data frame
with two columns: "xintercept" with the values, and "lake"
with the name of each lake in df. Default to NULL to not
add a line.}

\item{probs}{exceedance probabilities (e.g., 10 for 10\%) to identify with
a horizontal line}

\item{title}{string to use for title of plot, defaults to "".}

\item{text_size}{size of text, defaults to 12}

\item{vline_color}{color of vertical line, defaults to a red ("#c00000")}

\item{hist_color}{color of histogram bins, defaults to "grey80"}

\item{density_color}{color of density curve with alpha = 0.2, defaults to a
red ("#c00000")}

\item{line_size}{line size, default to 1}

\item{pfl_is_long}{defaults to TRUE to force the y-limits of plainfield lake
to the same as long lake.}

\item{labels}{logical defaults to TRUE to label exceedance probabilities (if displayed)}

\item{lakes}{vector of lakes to include in plot. Defaults to c("Pleasant",
"Long", "Plainfield")}
}
\value{
plot_obj, a plot with the distribution(s) of estimated lake elevation
}
\description{
Given a data frame with lake levels for a single lake or multiple lakes, plot
the distribution (density) of estimated lake levels. If data frame has
multiple lakes, will display as facet plots. Options to add xintercepts.
}
