% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lake_contours.R
\name{plot_lake_contours}
\alias{plot_lake_contours}
\title{Draw contours for a given lake}
\usage{
plot_lake_contours(
  lake,
  lake_levels,
  color_legend_title = "Exceedance Probability (\%)",
  color_labels = c(10, 50, 90),
  color_values = c("#a80000", "#ffab00", "#004da8"),
  legend_position = "top",
  title_name = NULL,
  text_size = 10
)
}
\arguments{
\item{lake}{name of lake to analyze, e.g., "Long"}

\item{lake_levels}{a single value or vector of lake elevations (mamsl) to
draw contors for.}

\item{color_legend_title}{legend title for colors.}

\item{color_labels}{a vector corresponding to lake_levels with the name to
use for each level in the legend. Defaults to
c(10, 50, 90) for the 10\%, 50\%, and 90\% exceedance
probabilities.}

\item{color_values}{a vector corresponding to lake_levels with the color to
use for each level in the legend. Defaults to
c("#a80000", "#ffab00", "#004da8") for Tuscan Red
(#a80000), Electron Gold (#ffab00), and Ultra Blue
(#004da8). Note that the order of this vector does not
necessarily correspond with the order of lake_levels or
color_labels, it may take some trial and error to find
the right order for the visualization you have in mind.}

\item{legend_position}{position of legend, defaults to "top", set to "none"
to remove legend (note: still need to include
color_name, color_labels, and color_values even if
leave the legend off the returned plot).}

\item{title_name}{string to use for plot title}

\item{text_size}{font size for text in plot}
}
\value{
plot_obj, a plot with the contours of the lake.
}
\description{
Given a) a single lake level or a vector of lake levels as well as b) the
name of the lake of interest, this loads the raster associated with the lake
and draws contours at the desired levels. Note that by default, each contour
is color coded. To turn this off, set color_values to the all the same color
(e.g., c("black", "black", "black")) and set legend_position to "none".
}
