% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_magnitude.R
\name{plot_magnitude}
\alias{plot_magnitude}
\title{Plot distribution of inputed lake levels}
\usage{
plot_magnitude(
  df,
  xintercept = NULL,
  exceedance = NULL,
  title = "",
  text_size = 12,
  vline_color = "#c00000",
  hist_color = "grey80",
  density_color = "#c00000",
  line_size = 1
)
}
\arguments{
\item{df}{a data frame with columns for "lake", "date", "level_obs", and
"level_pred".}

\item{xintercept}{either a single value (if only one lake) or a data frame
with two columns: "xintercept" with the values, and "lake"
with the name of each lake in df. Default to NULL to not
add a line.}

\item{title}{string to use for title of plot, defaults to "".}

\item{text_size}{size of text, defaults to 12}

\item{vline_color}{color of vertical line, defaults to a red ("#c00000")}

\item{hist_color}{color of histogram bins, defaults to "grey80"}

\item{density_color}{color of density curve with alpha = 0.2, defaults to a
red ("#c00000")}

\item{line_size}{line size, default to 1}
}
\value{
plot_obj, a plot with the distribution(s) of estimated lake elevation
}
\description{
Given a data frame with lake levels for a single lake or multiple lakes, plot
the distribution (density) of estimated lake levels. If data frame has
multiple lakes, will display as facet plots. Options to add xintercepts.
}
