% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_plant_populations.R
\name{plot_plant_populations}
\alias{plot_plant_populations}
\title{Fassett's Locoweed populations}
\usage{
plot_plant_populations()
}
\value{
nothing
}
\description{
Plots and saves fassett's locoweed populations
}
