% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rate_change.R
\name{plot_rate_change}
\alias{plot_rate_change}
\title{Plot rate of change}
\usage{
plot_rate_change(
  df,
  lakes = c("Pleasant", "Long", "Plainfield"),
  lag_months = 1,
  show_zero = TRUE,
  max_rate = NULL,
  title = "",
  text_size = 12,
  vline_color = "black",
  hist_color = "grey80",
  density_color = "#c00000",
  line_size = 1
)
}
\arguments{
\item{df}{a data frame with columns for "lake", "date", "level_obs", and
"level_pred" that matches desired time series to analyze}

\item{lakes}{name of lakes to analyze}

\item{lag_months}{number of months to analyze rise/fall over, defaults to 1}

\item{show_zero}{logical defaults to FALSE. If true, plots vertical line at
median number of months.}

\item{max_rate}{default to NULL, otherwise used to set max x limits}

\item{title}{string to use for title of plot, defaults to "".}

\item{text_size}{size of text, defaults to 12}

\item{vline_color}{color of vertical line, defaults to a red ("#c00000")}

\item{hist_color}{color of histogram bins, defaults to "grey80"}

\item{density_color}{color of density curve with alpha = 0.2, defaults to a
red ("#c00000")}

\item{line_size}{line size, defaults to 1}
}
\value{
plot_obj, a plot with the imputed and observed lake levels.
}
\description{
Given a data frame with lake levels for a single lake or multiple lakes, plot
the typical rate of change over a given time interval. If data frame has
multiple lakes, will display as facet plots.
}
