% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_short_v_long.R
\name{plot_short_v_long}
\alias{plot_short_v_long}
\title{Plot bias, precision, and accuracy of metrics}
\usage{
plot_short_v_long(
  df_short,
  df_long,
  metric_name,
  variable_breaks = "sort",
  variable_labels,
  value_title = "",
  convert_units = "",
  point_size = 0.5,
  text_size = 10,
  rotate_xticks = FALSE
)
}
\arguments{
\item{df_short}{data frame with the median/sd for hydrologic metrics w/a
short time series}

\item{df_long}{data frame with the median/sd for hydrologic metrics w/a
long time series}

\item{metric_name}{name of metric to display on this plot}

\item{variable_breaks}{vector with values to use for variable breaks in color
legend. Defaults to "sort" to convert unique options
in "variable" column to numeric values and sort in
ascending order.}

\item{variable_labels}{vector with strings to use for variable labels in
color legend. Defaults to "breaks" to indicate use
same values as breaks. Can also be "percent" to add a
percent symbol to values in variable_breaks.}

\item{value_title}{label for y-axis values}

\item{convert_units}{logical defulats to "". If "metersTOft", converts values
from meters to feet.}

\item{point_size}{size of points, defaults to 0.5,}

\item{text_size}{size of text on plots. Defaults to 12.}

\item{rotate_xticks}{defaults to FALSE. If TRUE, rotates x tick lables
90 degrees}
}
\description{
Creates a facet plot with lakes faceted horizontally and error estimates
faceted vertically.
}
