% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_veg_outlines.R
\name{plot_veg_outlines}
\alias{plot_veg_outlines}
\title{Draw contours for vegetation classes for given lake/elevation}
\usage{
plot_veg_outlines(
  lake,
  lake_level,
  deepest = data.frame(upland = -1, inland_beach = 1.6, emergent = 7, floating = 10),
  shallowest = data.frame(floating = 5, submergent = 1.6, emergent = -1, inland_beach =
    -1.6),
  contour_interval = 0.05,
  depth_to_meters = TRUE,
  color_values = c("#B15928", "#FDBF6F", "#B2DF8A", "#33A02C", "#A6CEE3", "#1F78B4",
    "#6A3D9A"),
  line_size = 1,
  legend_pos = "top",
  text_size = 12
)
}
\arguments{
\item{lake}{name of lake to analyze, e.g., "Long". Must be present in both
CSLSlevels::csls_levels and CSLSdata::lake_levels.}

\item{lake_level}{lake elevation (m) to use for classifying vegetation extents}

\item{deepest}{list with deepest depth each vegetation class can tolerate
(ft).}

\item{shallowest}{list with shallowest depth each vegetation class can
tolerate (ft). If depth_to_meters is TRUE, these will be
converted to meters. If shallowest not provided for
"upland", assumed to be the negative max depth of the lake.}

\item{contour_interval}{interval of contours to use when creating
relationship, same units as desired transition depths
data frame. Defaults to 0.05.}

\item{depth_to_meters}{defaults to TRUE, indicating transition depths will be
in meters and deepest and shallowest should be
converted from feet to meters.}

\item{color_values}{string with hex codes to use for vegetation classes, from
driest to wettest.}

\item{line_size}{line size for shoreline, defaults to 1.}

\item{legend_pos}{legend position. Defaults to "top" but can also be "none",
"right", etc.}

\item{text_size}{size of legend text, defaults to 12.}
}
\value{
plot_obj, a plot with the areal extents of vegetation classes at the
given lake level, with the lake level denoted with a black outline.
}
\description{
Creates plot with areal extents of vegetation classes at given lake level,
with the lake level denoted with a black outline.
}
