% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_lake_levels.R
\name{predict_lake_levels}
\alias{predict_lake_levels}
\title{Predict lake levels w/BSTS method}
\usage{
predict_lake_levels(
  study_lakes,
  lake_name = "Long Lake",
  start_date = "1938-01-01",
  seed = 2017
)
}
\arguments{
\item{study_lakes}{input data frame for BSTS calcs. Comes from
"//dnr/programs/DG/DG_Projects/Water Use/WU_Central_Sands_Study/data_mgmt/hist_lake_levs/study_lakes_pred_levs_20200302.csv"}

\item{lake_name}{name of lake to process. Defaults to "Long Lake".}

\item{start_date}{first date to include in estimated lake level timeseries.
Defaults to "1938-01-01".}

\item{seed}{randomization seed for bsts. Defaults to 2017.}
}
\description{
Function contains Bob Smail's code for predicting lake levels using Bayseian
Structural Time Series.
}
