% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_lake_level_functions.R
\name{rank_lake_levels}
\alias{rank_lake_levels}
\title{Rank lake levels}
\usage{
rank_lake_levels(levels, window_start = 1905, window_end = 2018,
  lake_names)
}
\arguments{
\item{levels}{a data frame with the lake name (lake_name), month and year of
prediction (obs_mo), and predicted lake level (lake_lev_pred)}

\item{window_start}{year to start considering lake level data (inclusive),
defaults to 1905}

\item{window_end}{year to stop considering lake level data (inclusive),
defaults to 2018}

\item{lake_names}{name of all lakes in order of desired factor levels.}
}
\value{
ranked, a data frame with the lake name (lake), observation month and
        year (obs_mo), lake level (level), and exceedance probability (prob).
}
\description{
Given a timeseries of lake levels and a specified time period this function
ranks lake levels by lake and calculates lake level exceedance probability
curve for each lake.
}
