% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_lake_levels_random.R
\name{rank_lake_levels_random}
\alias{rank_lake_levels_random}
\title{Rank lake levels for a random subset of years}
\usage{
rank_lake_levels_random(
  levels,
  nyears,
  ts_start = 1905,
  ts_end = 2018,
  lake_names = c("Pleasant Lake", "Long Lake", "Plainfield Lake")
)
}
\arguments{
\item{levels}{a data frame with the lake name (lake_name), month and year of
prediction (obs_mo), and predicted lake level (lake_lev_pred)}

\item{nyears}{number of years to sample}

\item{ts_start}{year to start considering lake level data (inclusive),
defaults to 1905}

\item{ts_end}{year to stop considering lake level data (inclusive), defaults
to 2018}

\item{lake_names}{name of all lakes in order of desired factor levels.}
}
\value{
ranked, a data frame with the lake name (lake), observation month and
year (obs_mo), lake level (level), and exceedance probability (prob).
}
\description{
Given a timeseries of lake levels and a specified number of years to sample,
this function ranks a random subset of lake levels by lake and calculates
lake level exceedance probability curve for each lake.
}
