% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_lake_levels_sensitivity.R
\name{rank_lake_levels_sensitivity}
\alias{rank_lake_levels_sensitivity}
\title{Rank lake levels for multiple windows of analysis}
\usage{
rank_lake_levels_sensitivity(
  levels,
  years,
  types = "forward",
  windows = 0,
  start_year = 1905,
  end_year = 2018
)
}
\arguments{
\item{levels}{a data frame with the lake name (lake_name), month and year of
prediction (obs_mo), and predicted lake level (lake_lev_pred)}

\item{years}{vector of analysis end years to stop considering lake level data
(inclusive)}

\item{types}{type of looping to do, defaults to "forward". Also "backward".}

\item{windows}{length of window of analysis. Defaults to zero to indicate use
entire record.}

\item{start_year}{start year of lake level data}

\item{end_year}{end year of lake level data}
}
\value{
ranked, a data frame with the lake name (lake), observation month and
year (obs_mo), lake level (level), exceedance probability (prob),
and end year of analysis (end_year).
}
\description{
Given a timeseries of lake levels and a specified series of years, this
function ranks lake levels by lake for windows with the specified end years
and calculates lake level exceedance probability.
}
