% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_levels_at_probs.R
\name{sensitivity_levels_at_probs}
\alias{sensitivity_levels_at_probs}
\title{Find lake level at given exceedance probability for multiple windows of analyses}
\usage{
sensitivity_levels_at_probs(
  levels,
  years,
  probs = c(10, 50, 90),
  types = "forward",
  windows = 0,
  start_ts = 1905,
  end_ts = 2018
)
}
\arguments{
\item{levels}{a data frame with the lake name (lake_name), month and year of
prediction (obs_mo), and predicted lake level (lake_lev_pred)}

\item{years}{the end years of the analysis window used to calculate lake level
exceedance probabilities.}

\item{probs}{the exceedance probabilities of interest (percent), defaults to
c(10, 50, 90)}

\item{types}{type of looping to do, defaults to "forward". Also "backward".}

\item{windows}{length of window of analysis. Defaults to zero to indicate use
entire record.}

\item{start_ts}{start year of lake level data}

\item{end_ts}{end year of lake level data}
}
\value{
df, a data frame with the following columns:
\item{lake}{the name of the lake ("Pleasant", "Long", or "Plainfield)}
\item{season}{the season (e.g., annual, spring, summer, winter, or fall)}
\item{prob}{the exceedance probability}
\item{level}{the lake level associated with the given exceedance probability
for the given lake, season, and window of analysis}
\item{year}{year unique to this window of anlaysis (start year or end year)}
\item{sim_no}{id associated with this window of analysis}
}
\description{
Given a dataset with lake levels and exceedance probabilities, this function
interpolates the lake level associated with the provided exceedance
probability for each lake in the dataset.
}
