% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_probs_of_levels.R
\name{sensitivity_probs_of_levels}
\alias{sensitivity_probs_of_levels}
\title{Calculate sensitivity of exceedance probabilities to window of anlaysis}
\usage{
sensitivity_probs_of_levels(
  levels,
  years,
  types = "forward",
  windows = 0,
  start_ts = 1905,
  end_ts = 2018
)
}
\arguments{
\item{levels}{a data frame with the following columns:
\itemize{
\item \strong{lake:} the name of the lake ("Pleasant", "Long", or "Plainfield")
\item \strong{date:} the month an year of prediction in datetime (POSIXct) format
\item \strong{level_pred:} predicted lake level, in meters above mean sea level
}}

\item{years}{vector of years to loop through (either start years or end years,
depending on "types").}

\item{types}{type of looping to do, defaults to "forward". Also "backward".}

\item{windows}{length of window of analysis. Defaults to zero to indicate use
entire record.}

\item{start_ts}{start year of lake level data}

\item{end_ts}{end year of lake level data}
}
\value{
ranked, a data frame with the following columns:
\item{lake}{the name of the lake ("Pleasant", "Long", or "Plainfield)}
\item{season}{the season of the prediction}
\item{date}{the month and year of prediction in datetime (POSIXct) format}
\item{level}{the predicted lake level (meters above mean sea level)}
\item{prob}{the probability of a lake level that equals or exceeds the
predicted level (percent)}
\item{year}{year unique to this window of anlaysis (start year or end year)}
\item{sim_no}{id associated with this window of analysis}
}
\description{
Given a timeseries of lake levels and a specified series of years, this
function ranks lake levels by lake for multiple analysis windows and
calculates lake level exceedance probability.
}
