% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_range.R
\name{specify_range}
\alias{specify_range}
\title{Create data frame with equal ranges for all groups}
\usage{
specify_range(
  df,
  group_col,
  value_col,
  tick_precision = NA,
  pfl_is_long = FALSE
)
}
\arguments{
\item{df}{a data frame with columns for the group and corresponding values}

\item{group_col}{name of the group column}

\item{value_col}{name of the values column}

\item{tick_precision}{round range to nearest X, inclusive (e.g., 0.1, 0.5, etc.)}

\item{pfl_is_long}{logical defaults to FALSE to maintain separate ranges for
Long and Plainfield lakes. If TRUE, forces plainfield to
use range for long.}
}
\value{
range, a data frame with a column for "group" and another column for
"value" (real names: group_col and value_col) with min and max values
for each group.
}
\description{
Given a data frame with values corresponding to different groups, defines a
data frame "range" with min and max values for each group such that the real
data can be centered around the midpoint but all subplots cover the same
range (1 unit on plot A = 1 unit on plot B).
}
