% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo_functions.R
\name{summarise_monte_carlo}
\alias{summarise_monte_carlo}
\title{Summarize Monte Carlo Analysis of Lake Levels}
\usage{
summarise_monte_carlo(monte_carlo)
}
\arguments{
\item{monte_carlo}{a data frame with the lake name (lake), exceedance
probabilities (prob), associated lake levels (level), and
number of random years selected for analysis (year) in
monte carlo simulations of lake level exceedance curves.}
}
\value{
monte_carlo_summary, a summary of the input data frame with the mean
        and standard deviation of lake levels for each year.
}
\description{
This function summarizes the mean and sd of lake levels resulting from a
monte carlo analysis of lake level exceedance probabilities.
}
