% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_summary.R
\docType{data}
\name{ts_summary}
\alias{ts_summary}
\title{Dataset: Summary of hydrologic metric fit with full timeseries for different timeseries lengths}
\format{
A data frame with the following columns.
\describe{
\item{lake}{name of lake, e.g., Pleasant, Long, Plainfield, Devils}
\item{metric}{name of hydrologic metrics, e.g. median_level, cv_rise_rate}
\item{variable}{name of variation on metrics, e.g. 10 for 10\% exceedance
level}
\item{nyear}{number of years randomly sampled in this simulation}
\item{fit}{type of fit analyzed, PBIAS, CV, or RMSE}
\item{value}{value of fit}
}
}
\usage{
data(ts_summary)
}
\description{
Dataset: Summary of hydrologic metric fit with full timeseries for different timeseries lengths
}
\keyword{datasets}
