% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_Kennard_metrics.R
\docType{data}
\name{which_Kennard_metrics}
\alias{which_Kennard_metrics}
\title{A character vector of potential metrics for Kennard_metrics()}
\format{A vector of abbreviations of metrics to help with the Kennard_metrics() function}
\usage{
which_Kennard_metrics
}
\description{
A character vector of potential metrics for Kennard_metrics()
}
\examples{
Kennard_metrics(metrics = which_Kennard_metrics)

}
\keyword{datasets}
