% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_exceedances.R
\name{calculate_exceedances}
\alias{calculate_exceedances}
\title{Calculate exceedance levels}
\usage{
calculate_exceedances(
  df,
  probs = c(10, 25, 50, 75, 90),
  departures = NULL,
  melted = TRUE
)
}
\arguments{
\item{df}{a data frame with a "lake" and a "level" column}

\item{probs}{a vector with all exceedance probabilities to calculate.
Defaults to c(10, 25, 50, 75, 90)}

\item{departures}{optional vector with departures from median to evaluate
probability of occurrence. Defaults to NULL to trigger exceedance
probability calculations instead.}

\item{melted}{logical defaults to true to indicate should melt data frame to
just 3 columns (lake, variable, value). Otherwise, keeps as
data frame with one column per exceedance probability.}
}
\value{
exceeds, a data frame with the names of all lakes and corresponding
exceedance levels
}
\description{
Given a data frame with a "lake" and "level" column, calculates levels
associated with desired exceedance probabilities.
}
