% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_metrics.R
\name{calculate_metrics}
\alias{calculate_metrics}
\title{Calculate hydrologic metrics evaluated in the CSLS}
\usage{
calculate_metrics(
  df = CSLSlevels::csls_levels,
  col_name = "level_pred",
  start_date = NULL,
  end_date = NULL,
  metrics = c("median_level", "cv_level", "exceedance_level", "exceedance_range",
    "median_dur", "cv_dur", "median_rise_rate", "cv_rise_rate", "median_fall_rate",
    "cv_fall_rate")
)
}
\arguments{
\item{df}{data frame to use, defaults "CSLSlevels::csls_levels".}

\item{col_name}{name of column with lake level values to use. Defaults to
"level_pred".}

\item{start_date}{Earliest date to use in calculations in datetime format.
Defaults to NULL to use entire range of dates in df.}

\item{end_date}{Latest date to use in calculations in datetime format.
Defaults to NULL to use entire range of dates in df.}

\item{metrics}{a list of which metrics to use. Defaults to all of them
c("median_level", "cv_level", "exceedance_level",
"exceedance_range", "median_dur", "cv_dur",
"median_rise_rate", "cv_rise_rate", "median_fall_rate",
"cv_fall_rate").}
}
\value{
summary, a data frame with the following columns:
\item{lake}{name of lake, character}
\item{metric}{name of metric, corresponds with values in inputted "metrics"
argument, character}
\item{variable}{identifier for different types of metric. For example, the
metric "median_rise_rate" will have variables of "1", "3",
and "12" associated with it indicating the median rate over
1 month, 3 months, and 12 months, character}
\item{value}{calculated value of the metric, numeric}
}
\description{
Calculate hydrologic metrics evaluated in the CSLS
}
