% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lake_depths.R
\name{plot_lake_depths}
\alias{plot_lake_depths}
\title{Draw filled depth contours for a given lake}
\usage{
plot_lake_depths(
  lake,
  lake_level,
  relative_depth_fill = TRUE,
  legend_position = "right",
  title_name = NULL,
  fill_name = "Depth (m)",
  text_size = 10,
  lat_long_off = TRUE,
  scale_bar = "bl",
  scale_bar_on = TRUE
)
}
\arguments{
\item{lake}{name of lake to analyze, e.g., "Long"}

\item{lake_level}{a single lake elevations (mamsl) to draw depth contors for.}

\item{relative_depth_fill}{logical defaults to TRUE to scale limits of
fill color to maximum observed depth.}

\item{legend_position}{position of legend, defaults to "top", set to "none"
to remove legend (note: still need to include
color_name, color_labels, and color_values even if
leave the legend off the returned plot).}

\item{title_name}{string to use for plot title}

\item{fill_name}{string to use for fill legend title, defaults to "Depth (m)".}

\item{text_size}{font size for text in plot}

\item{lat_long_off}{logical for whether to include lat long axis labels.
Defaults to TRUE to remove axis tick marks and labels.}

\item{scale_bar}{location of annotation scale bar, defaults to "bl" for
bottom left ("tl" = top left, etc.)}

\item{scale_bar_on}{logical defaults to true to display scale bar.}
}
\value{
plot_obj, a plot with the contours of the lake.
}
\description{
Given a single lake level and the name of the lake of interest, this loads
the raster associated with the lake and draws filled depth contours.
Optionally scales the limits of the fill to the maximum observed depth, or
relative to the maximum depth at the given lake level.
}
