% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_levels.R
\name{plot_levels}
\alias{plot_levels}
\title{Plot inputed lake levels and observations}
\usage{
plot_levels(
  df,
  lakes = c("Pleasant", "Long", "Plainfield"),
  yintercept = NULL,
  exceedance = NULL,
  depth_axis = TRUE,
  title = "",
  legend_pos = c(0.15, 0.95),
  text_size = 12,
  color_vals = c("grey70", "black"),
  line_size = 1,
  point_size = 3
)
}
\arguments{
\item{df}{a data frame with columns for "lake", "date", "level_obs", and
"level_pred".}

\item{lakes}{names of lakes in df in order to display. Defaults to
c("Pleasant", "Long", "Plainfield")}

\item{yintercept}{either a single value (if only one lake) or a data frame
with two columns: "yintercept" with the values, and "lake"
with the name of each lake in df. Default to NULL to not
add a line.}

\item{exceedance}{exceedance probability (e.g., 10 for 10\%) to identify with
a horizontal line}

\item{depth_axis}{logical defaults to TRUE to add a second axis with lake
depth if there's only one lake in the df}

\item{title}{string to use for title of plot, defaults to "".}

\item{legend_pos}{position of legend. Defaults to c(0.15, 0.95) but can also
be "top", "right", etc.}

\item{text_size}{size of text, defaults to 12}

\item{color_vals}{colors to use for line and points, defaults to c("grey70",
"black") for grey lines, black points}

\item{line_size}{line size, defaults to 1}

\item{point_size}{point size, defaults to 3}
}
\value{
plot_obj, a plot with the imputed and observed lake levels.
}
\description{
Given a data frame with lake levels for a single lake or multiple lakes, plot
the estimated and observed lake levels. If data frame has multiple lakes,
will display as facet plots. Options to add yintercepts. Also option to add
second axis with lake depth, if only plotting one lake.
}
