% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_magnitude.R
\name{plot_magnitude}
\alias{plot_magnitude}
\title{Plot distribution of inputed lake levels}
\usage{
plot_magnitude(
  df,
  lakes = c("Pleasant", "Long", "Plainfield"),
  convert_to_ft = TRUE,
  title = "",
  text_size = 12,
  pfl_is_long = TRUE,
  force_range = TRUE
)
}
\arguments{
\item{df}{a data frame with columns for "lake", "date", and "level".}

\item{lakes}{vector of lakes to include in plot. Defaults to c("Pleasant",
"Long", "Plainfield")}

\item{convert_to_ft}{defaults to TRUE to convert lake levels from meters to ft}

\item{title}{string to use for title of plot, defaults to "".}

\item{text_size}{size of text, defaults to 12}

\item{pfl_is_long}{defaults to TRUE to force the y-limits of plainfield lake to the same as long lake.}

\item{force_range}{defaults to TRUE to force y-limits of all subplots to be the same.}
}
\value{
plot_obj, a plot with the distribution(s) of estimated lake elevation
}
\description{
Given a data frame with lake levels for a single lake or multiple lakes, plot
the distribution (density) of estimated lake levels. If data frame has
multiple lakes, will display as facet plots. Options to add xintercepts.
}
