% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sensitivity.R
\name{plot_sensitivity}
\alias{plot_sensitivity}
\title{Plot bias, precision, and accuracy of metrics}
\usage{
plot_sensitivity(
  df,
  metric_name,
  metric_title = "",
  variable_title = "",
  variable_breaks = "sort",
  variable_labels,
  variable_colors = NULL,
  text_size = 12,
  vline = 38
)
}
\arguments{
\item{df}{data frame with the following columns:
\itemize{
\item{lake}{name of lake, e.g., Pleasant, Long, Plainfield, Devils}
\item{metric}{name of hydrologic metrics, e.g. median_level, cv_rise_rate}
\item{variable}{name of variation on metrics, e.g. 10 for 10\% exceedance
level}
\item{nyear}{number of years used to calculate metric}
\item{fit}{type of fit measurement, PBIAS, CV, or RMSE}
\item{value}{value of fit measurement}
}}

\item{metric_name}{name of metric to display on this plot}

\item{metric_title}{string to use for plot title, defaults to "" for no title}

\item{variable_title}{string to use for color legend, defaults to "" for no
title}

\item{variable_breaks}{vector with values to use for variable breaks in color
legend. Defaults to "sort" to convert unique options
in "variable" column to numeric values and sort in
ascending order.}

\item{variable_labels}{vector with strings to use for variable labels in
color legend. Defaults to "breaks" to indicate use
same values as breaks. Can also be "percent" to add a
percent symbol to values in variable_breaks.}

\item{variable_colors}{vector with hex codes of colors corresponding to
variable_breaks and variable_lables.}

\item{text_size}{size of text on plots. Defaults to 12.}

\item{vline}{optional value at which to draw a dashed vertical line. Defaults
to 38 (i.e., nyears from 1981-2018), but can be set to NULL for
no line.}
}
\description{
Creates a facet plot with lakes faceted horizontally and error estimates
faceted vertically.
}
