% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timing.R
\name{plot_timing}
\alias{plot_timing}
\title{Plot monthly timeseries}
\usage{
plot_timing(
  df,
  title = "",
  text_size = 12,
  force_pfl = TRUE,
  npretty_breaks = NULL
)
}
\arguments{
\item{df}{a data frame with columns for "lake", "date", "level_obs", and
"level_pred".}

\item{title}{string to use for title of plot, defaults to "".}

\item{text_size}{size of text, defaults to 12}

\item{force_pfl}{defaults to TRUE to force the y-limits of plainfield lake to
the same as long lake.}

\item{npretty_breaks}{defaults to NULL, set to 3 to limit Long, Plainfield,
and Pleasant to integers (or play around, may be different with different
periods of record)}
}
\value{
plot_obj, a plot with the imputed and observed lake levels.
}
\description{
Given a data frame with lake levels for a single lake or multiple lakes, plot
the estimated and observed lake levels. If data frame has multiple lakes,
will display as facet plots. Options to add yintercepts. Also option to add
second axis with lake depth, if only plotting one lake.
}
